+++ src/Resources/WorkflowResource/Pages/CreateWorkflow.php
@@ -2,10 +2,14 @@
 
 namespace Xentixar\WorkflowManager\Resources\WorkflowResource\Pages;
 
-use Xentixar\WorkflowManager\Resources\WorkflowResource;
-use Filament\Actions;
+use Exception;
+use Filament\Notifications\Notification;
 use Filament\Resources\Pages\CreateRecord;
+use Illuminate\Database\Eloquent\Relations\Relation;
+use Xentixar\WorkflowManager\Models\Workflow;
 use Xentixar\WorkflowManager\Models\WorkflowState;
+use Xentixar\WorkflowManager\Resources\WorkflowResource;
+use Xentixar\WorkflowManager\Support\Helper;
 
 class CreateWorkflow extends CreateRecord
 {
@@ -20,14 +24,26 @@ class CreateWorkflow extends CreateRecord
 
     protected function afterCreate(): void
     {
-        $states = (new $this->record->model_class())->getStates() ?: [];
-        foreach ($states as $key => $state) {
-            WorkflowState::query()
-                ->create([
-                    'workflow_id' => $this->record->id,
-                    'state' => $key,
-                    'label' => $state ?? null,
-                ]);
+        /** @var Workflow $record */
+        $record = $this->record;
+
+        try {
+            $enumClass = ($record->model_class)::getStates() ?: ''; // @phpstan-ignore-line
+            $states = Helper::getStatesFromEnum($enumClass);
+            foreach ($states as $key => $state) {
+                WorkflowState::query()
+                    ->create([
+                        'workflow_id' => $record->id,
+                        'state' => $key,
+                        'label' => $state,
+                    ]);
+            }
+        } catch (Exception $exception) {
+            Notification::make('error')
+                ->title('Error')
+                ->body($exception->getMessage())
+                ->danger()
+                ->send();
         }
     }
 }