+++ src/Support/Helper.php
@@ -3,14 +3,17 @@
 namespace Xentixar\WorkflowManager\Support;
 
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
+use Throwable;
+use Xentixar\WorkflowManager\Contracts\WorkflowsContract;
 
 final class Helper
 {
     /**
      * Get available models in app that implement the Workflows interface.
-     * 
-     * @return array $models
+     *
+     * @return array<string, string> $models
      */
     public static function getAvailableModels(): array
     {
@@ -19,19 +22,50 @@ final class Helper
         $modelFiles = File::allFiles($modelsPath);
 
         foreach ($modelFiles as $modelFile) {
-            $class = 'App\\Models\\' . Str::replaceLast('.php', '', $modelFile->getFilename());
+            $class = 'App\\Models\\'.Str::replaceLast('.php', '', $modelFile->getFilename());
 
-            if (!class_exists($class)) {
+            if (! class_exists($class)) {
                 continue;
             }
 
-            $implementedInterfaces = class_implements($class);
+            try {
+                $implementedInterfaces = class_implements($class);
+                $morphClass = (new $class)->getMorphClass(); // @phpstan-ignore-line
 
-            if (in_array(\Xentixar\WorkflowManager\Contracts\Workflows::class, $implementedInterfaces)) {
-                $models[$class] = $class;
+                if (in_array(WorkflowsContract::class, $implementedInterfaces)) {
+                    $models[$morphClass] = class_basename($class);
+                }
+            } catch (Throwable $e) {
+                Log::error('Error while checking model interfaces: '.$e->getMessage());
+
+                continue;
             }
         }
 
         return $models;
+    }
+
+    /**
+     * Get available states from enum class
+     *
+     * @return array<string, string>
+     */
+    public static function getStatesFromEnum(string $enumClass): array
+    {
+        if (! class_exists($enumClass)) {
+            return [];
+        }
+
+        $states = [];
+
+        foreach ($enumClass::cases() as $state) {
+            if (method_exists($state, 'getLabel')) {
+                $label = $state->getLabel();
+            }
+
+            $states[$state->value] = $label ?? $state->value;
+        }
+
+        return $states;
     }
 }